Add dataset downloader, first draft
 - multi-greenlet downloading from all servers the host resolves to
 - unpacking of GRIBS to huge memory mapped data file
 - dumping of GRIB records to "grib mirror" (a quarter of the size of
   the data downloaded by excluding variables we don't need)

Developing this fast was forced by the OPeNDAP server cound being
reduced to one, making its performance too slow for the old predictor. I
intend to jury-rig tawhiri's downloader into the old predictor while the
rest of tawhiri is finished, in order to solve the immediate problem.
I'll then come back and add the following things, which are missing:

 - docstrings
 - documentation
 - tests