setup.py: add dependency link for ruaumoko
As it stands, trying to git clone the repo and perform the standard
``pip install -r requirements.txt`` or ``python setup.py develop`` dance
will fail unless you have ruaumoko[1] installed. If it is not installed,
setuptools attempts to download it from PyPI which fails.

Add a ``dependency_links`` parameter to the call to ``setup`` which
specifies where to fetch the code for ruaumoko.

[1] https://github.com/cuspaceflight/ruaumoko/