travis: add configuration and empty test suite
This copies the Travis configuration from ruaumoko and adds an empty
test suite. Somewhat confusingly we now have ``tests`` and ``testing``
in the tree. I did consider re-using the existing ``testing`` directory
but the tiles there are not set up to be a traditional test suite and
so, for the moment, I explicitly exclude them.

I intend to move files from ``testing`` into ``tests`` over time and, if
any are left, rename that directory to ``examples`` or ``scripts`` which
probably more accurately reflects what purpose it will have then.