download: use six's reraise function
Replace uses of three-argument raise with six's reraise function. On
Python 2 this is exactly equivalent and on Python 3 it Does The Right
Thing(TM). The downloader is Python 2-only for other reasons but using
this function at least makes it *syntactically* correct in Python 3.

Having the file parse-able with Python 3 means that sphinx does not barf
when building the documentation with a Python 3 virtualenv.